Arrumando npm
